+++ changelog.md
@@ -15,6 +15,7 @@ Bug Fixes:
 * Add enable_pager to the config file (Thanks: [Frederic Aoustin]).
 * Mark `test_sql_output` as a dbtest (Thanks: [Dick Marinus]).
 * Don't crash if the log/history file directories don't exist (Thanks: [Thomas Roten]).
+* Unquote dsn username and password (Thanks: [Dick Marinus]).
 
 1.16.0:
 =======